+++ src/main/java/org/azfar/vatcalculator/exporter/Exporter.java
@@ -1,5 +1,7 @@
 package org.azfar.vatcalculator.exporter;
 
+import java.math.BigDecimal;
+
 public interface Exporter {
-    void export(double vatAmount, String filename);
+    void export(BigDecimal vatAmount, String filename);
 }
